makes topbar reusable
