+++ handraiser-app/src/components/mentor/MentorUi.js
@@ -31,6 +31,7 @@ import {
 import axios from "axios";
 
 import Tabs from "./Tabs/Tabs";
+import Topbar from "../reusables/Topbar"
 
 export default function Student() {
   let history = useHistory();
@@ -84,49 +85,10 @@ export default function Student() {
     }
   }, []);
 
-  if (state.user_type === 1) {
+  if (state.user_type === 4) {
     return (
       <React.Fragment>
-        <Nav>
-          <AppBar style={{ backgroundColor: "#372476" }}>
-            <Toolbar
-              style={{
-                display: "flex",
-                justifyContent: "space-between"
-              }}
-            >
-              <div style={{ display: "flex", alignItems: "center" }}>
-                <IconButton edge="start" aria-label="menu">
-                  <MenuIcon style={{ color: "white" }} />
-                </IconButton>
-                <Typography variant="h6">Handraiser Admin</Typography>
-              </div>
-              <div>
-                <IconButton
-                  aria-label="account of current user"
-                  edge="end"
-                  onClick={handleMenu}
-                  color="inherit"
-                >
-                  <AccountCircle style={{ fontSize: 40 }} />
-                </IconButton>
-              </div>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                  vertical: "top",
-                  horizontal: "right"
-                }}
-                open={open}
-                onClose={handleClose}
-              >
-                <MenuItem onClick={handleClose}>Profile</MenuItem>
-                <MenuItem onClick={handleClose}>Log Out</MenuItem>
-              </Menu>
-            </Toolbar>
-          </AppBar>
-        </Nav>
+        <Topbar/>
         <Div>
           <Queue>
             <Tabs />
