+++ handraiser-app/src/components/student/StudentUi.js
@@ -1,13 +1,5 @@
 import React, { useState, useEffect } from "react";
-import styled from "styled-components";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
-import IconButton from "@material-ui/core/IconButton";
-import MenuIcon from "@material-ui/icons/Menu";
-import AccountCircle from "@material-ui/icons/AccountCircle";
-import MenuItem from "@material-ui/core/MenuItem";
-import Menu from "@material-ui/core/Menu";
 import TextField from "@material-ui/core/TextField";
 import MoreVertIcon from "@material-ui/icons/MoreVert";
 import Swal from "sweetalert2";
@@ -32,6 +24,7 @@ import {
 import axios from "axios";
 
 import Tabs from "./Tabs/Tabs";
+import Topbar from "../reusables/Topbar";
 
 export default function Student() {
   let history = useHistory();
@@ -39,12 +32,11 @@ export default function Student() {
   const [state, setState] = useState({ user_type: "" });
   const open = Boolean(anchorEl);
   const [name, setName] = useState("");
+
   const handleMenu = event => {
     setAnchorEl(event.currentTarget);
   };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
+
   const sendMsg = evt => {
     evt.preventDefault();
     console.log(name);
@@ -88,46 +80,7 @@ export default function Student() {
   if (state.user_type === 3) {
     return (
       <React.Fragment>
-        <Nav>
-          <AppBar style={{ backgroundColor: "#372476" }}>
-            <Toolbar
-              style={{
-                display: "flex",
-                justifyContent: "space-between"
-              }}
-            >
-              <div style={{ display: "flex", alignItems: "center" }}>
-                <IconButton edge="start" aria-label="menu">
-                  <MenuIcon style={{ color: "white" }} />
-                </IconButton>
-                <Typography variant="h6">Handraiser Admin</Typography>
-              </div>
-              <div>
-                <IconButton
-                  aria-label="account of current user"
-                  edge="end"
-                  onClick={handleMenu}
-                  color="inherit"
-                >
-                  <AccountCircle style={{ fontSize: 40 }} />
-                </IconButton>
-              </div>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                  vertical: "top",
-                  horizontal: "right"
-                }}
-                open={open}
-                onClose={handleClose}
-              >
-                <MenuItem onClick={handleClose}>Profile</MenuItem>
-                <MenuItem onClick={handleClose}>Log Out</MenuItem>
-              </Menu>
-            </Toolbar>
-          </AppBar>
-        </Nav>
+        <Topbar />
         <Div>
           <Queue>
             <Tabs />
